requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.6.2
cffi==1.17.0
charset-normalizer==3.3.2
contourpy==1.2.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==43.0.0
cycler==0.12.1
defusedxml==0.8.0rc2
Django==4.2.9
django-cors-headers
django-debug-toolbar
django-templated-mail==1.1.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
drf-yasg==1.21.4
fonttools==4.53.1
frozendict==2.4.4
html5lib==1.1
idna==3.7
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
kiwisolver==1.4.5
lxml==5.2.2
MarkupSafe==2.1.5
multitasking==0.0.11
numpy==1.26.0
oauthlib==3.2.2
packaging==24.1
patsy==0.5.6
peewee==3.17.6
pillow==10.3.0
platformdirs==4.2.2
pycparser==2.22
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2024.1
psycopg2-binary==2.9.7
requests==2.32.3
requests-oauthlib==2.0.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
scipy==1.11.2
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
webencodings==0.5.1
